display usage on sevod-scraper.sh and additional info on readme

+++ assembly/src/script/sevod-scraper.sh
@@ -13,5 +13,14 @@ case $1 in
    rdfdump)   MAIN=org.semagrow.sevod.scraper.rdf.dump.RdfDumpScraper ;;
 esac
 
+if [ -e $MAIN ]
+then
+    echo "Usage:"
+    echo "./sevod-scraper.sh rdfdump [dump_file] [endpoint_url] [-s|p|o|j] [output_file]"
+    echo "./sevod-scraper.sh rdfdump [dump_file] [endpoint_url] [-s|p|o|j] [subjectTrieParameter] [objectTrieParameter] [output_file]"
+    echo "./sevod-scraper.sh cassandra [cassandra-ip-address] [cassandra-port] [cassandra-keyspace] [base] [output_file]"
+    exit
+fi
+
 shift
 $JAVA -cp $JARS $MAIN $*
